Protocol - GenerateTimestamp: provide old timestamp if regenerating
Note that the current timestamping/commiting scheme was not really
fault tolerant, or at least would require some strong synchronization
at servers.

Consider the following scenario:
- client requests a timestamp and receives it, the timestamp is valid
  for some limited amount of time (unused timestamps are GCed)
- client sends the commit request with updates using this timestamp
- connection breaks or server fails
- resending the updates with the same timestamp is safe (idempotent)
- however if timestamp got invalidated, it is much harder to ensure safety,
  since issuing a new timestamp requires servers to make sure
  that the updates are never committed using two different timestamps
  (concurrently, or sequentially)

To allow safe GC at server, the protocol is extended with optional
previousTimestamp field that specifies the previously used and rejected
timestamp. It allows servers to ensure that the old timestamp was not
used, GCed, and safely assign a new timestamp. This can be done without
synchronization on the critical path, and only by ensuring that there is
no concurrent GC on previousTimestamp when it is given as a paremeter.

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/concordant/sources/swiftcloud/java/SwiftCloud@310 049c2a90-bdf1-4f83-8191-3719d539f8e0