Big client refactor and related changes/fixes
Rationale:
Support anything different than SI, which requires in particular
dealing properly with transaction commiting to the store concurrently to
get requests (this issue was only partially solved till now)

Main complexity source:
Both locally and globally (on another client) there may be some
transactions depending on updates of commiting transaction; when they
both become visible, we must use the same timestamp for commiting
transaction.

Changes:
TxnHandleImpl was simplified and turned into  AbstractTxnHandle, with
SnapshotIsolationTxnHandle as the first implementation. The code dealing
with local transaction dependencies is now roughly self-contained in SwiftImpl.
Concurrency of SwiftImpl has improved - it can now commit a transaction
to the store and apply it with global timestamp in the cache while serving
a pending local transaction at the same time (yay if it only works!).
Some changes were required in the code dealing with operation dependencies
here and there.

Next steps:
- more tests and docs
- timeouts in get() request
- REPEATABLE_READS implementation
- notifications

git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/concordant/sources/swiftcloud/java/SwiftCloud@427 049c2a90-bdf1-4f83-8191-3719d539f8e0